Update geojson processing script to handle workarounds
Some of the state/district names have changed since Census data is collected
and we need to handle them while processing census data.

To handle this efficiently, create a workaround_data.csv in which we include
previous names and corrected names and the script takes care of the rest.

Only few of them are handled in this commit, rest need to be populated.

Also change some of the function signatures in geojson processing script to
follow logical order.

Signed-off-by: Arun Siluvery <arunk.siluvery@gmail.com>